fix: set DFX_VERSION and add dfx bin path to PATH when calling dfx proxy (#68)
When proxying to dfx:
- set DFX_VERSION so that dfx doesn't in turn forward to something else
- prepend the dfx bin path to PATH so that any calls to dfx in package.json or build scripts will go to the right place

All dfx binaries to date look at DFX_VERSION and the "dfx" field in dfx json.  If these indicate to use a different version of dfx, they forward to that version instead.  By setting DFX_VERSION, we ensure that dfx won't forward to a different version.

Previous installed versions of dfx were installed to /usr/local/bin.  While it's likely that the dfxvm bin directory is on the path, adding the dfx version bin directory to the path will ensure that any calls to dfx go to the right version.

Fixes https://dfinity.atlassian.net/browse/SDK-1362